requirements.txt
Pillow
numpy
ultralytics
easyocr
djangorestframework
